Committed IDView changes, added error display to VerificationRequestView
Right now, a lot of code is duplicated across IDView and
VerificationRequestView. Ideally, all the error message displaying logic would be
placed is a base class, and all views that need to display errors would
inherit from that class.

This is not directly possible here as IDView and
VerificationRequestView inherit from different base classes.